feat(U1): scaffolding — move recognizers_ja, add [bench] extras, rewrite Dockerfile
- git mv server/src/recognizers_ja.py packages/training/src/pleno_ner_training/recognizers_ja.py
- server/src/app.py + server/tests/conftest.py import updates (pleno_ner_training.recognizers_ja)
- packages/training/pyproject.toml [bench] extras (presidio, ginza, ja-ginza, numpy, scipy, ja_core_news_trf 3.8.0 wheel URL) + hatch.metadata.allow-direct-references
- server/pyproject.toml workspace dep on pleno-ner-training (tool.uv.sources)
- Dockerfile workspace-aware rewrite (COPY pyproject.toml + uv.lock + workspace members; CMD server.src.app:app); [bench] extras excluded by default since they live in optional-dependencies
- packages/training/tests/ scaffold (__init__.py, conftest.py)

Plan: docs/plans/2026-05-02-001-feat-ginza-presidio-baseline-measurement-plan.md (U1)
