+++ packages/training/tests/conftest.py
@@ -1,0 +1,5 @@
+"""Shared pytest fixtures for pleno-ner-training tests.
+
+Currently empty; per-test-file fixtures are defined locally. This file exists
+so that pytest discovers `packages/training/tests/` as a test root (U1 scaffold).
+"""
